small events refactor
